==> src/2_analysis/00-run-analyses.R <==


source("0-config.R")

source(here("src/2_analysis/01-primary-analysis.R"))
source(here("src/2_analysis/02-primary-PAF.R"))
source(here("src/2_analysis/03-secondary-analyses.R"))

source(here("src/2_analysis/03-SRMA-reference-analysis.R"))
source(here("src/2_analysis/03-SRMA-reference-analysis-unadjusted.R"))

source(here("src/2_analysis/03-POC-analysis-multinomial.R"))
source(here("src/2_analysis/03-POC-analysis-mort.R"))
source(here("src/2_analysis/03-POC-analysis-mort-multinomial.R"))
source(here("src/2_analysis/03-POC-analysis-tmle.R"))
